+++ Source/Core/CesiumTerrainProvider.js
@@ -85,7 +85,7 @@ define([
      *
      * // Terrain geometry near the surface of the globe is difficult to view when using NaturalEarthII imagery,
      * // unless the TerrainProvider provides additional lighting information to shade the terrain (as shown above).
-     * var imageryProvider = Cesium.createTileMapServiceImageryProvider({
+     * var imageryProvider = new Cesium.TileMapServiceImageryProvider({
      *        url : 'http://localhost:8080/Source/Assets/Textures/NaturalEarthII',
      *        fileExtension : 'jpg'
      *    });