+++ Specs/Scene/ImageryLayerSpec.js
@@ -11,7 +11,7 @@ defineSuite([
         'Renderer/TextureMinificationFilter',
         'Scene/ArcGisMapServerImageryProvider',
         'Scene/BingMapsImageryProvider',
-        'Scene/createTileMapServiceImageryProvider',
+        'Scene/TileMapServiceImageryProvider',
         'Scene/GlobeSurfaceTile',
         'Scene/Imagery',
         'Scene/ImageryLayerCollection',
@@ -36,7 +36,7 @@ defineSuite([
         TextureMinificationFilter,
         ArcGisMapServerImageryProvider,
         BingMapsImageryProvider,
-        createTileMapServiceImageryProvider,
+        TileMapServiceImageryProvider,
         GlobeSurfaceTile,
         Imagery,
         ImageryLayerCollection,
@@ -487,7 +487,7 @@ defineSuite([
 
     describe('createTileImagerySkeletons', function() {
         it('handles a base layer that does not cover the entire globe', function() {
-            var provider = createTileMapServiceImageryProvider({
+            var provider = new TileMapServiceImageryProvider({
                 url : 'Data/TMS/SmallArea'
             });
 
@@ -536,7 +536,7 @@ defineSuite([
                 url : 'Data/Images/Blue.png'
             });
 
-            var provider = createTileMapServiceImageryProvider({
+            var provider = new TileMapServiceImageryProvider({
                 url : 'Data/TMS/SmallArea'
             });
 
@@ -582,7 +582,7 @@ defineSuite([
                 url : 'Data/Images/Green4x4.png'
             });
 
-            var provider = createTileMapServiceImageryProvider({
+            var provider = new TileMapServiceImageryProvider({
                 url : 'Data/TMS/SmallArea'
             });