+++ Specs/Scene/OpenStreetMapImageryProviderSpec.js
@@ -1,5 +1,5 @@
 defineSuite([
-        'Scene/createOpenStreetMapImageryProvider',
+        'Scene/OpenStreetMapImageryProvider',
         'Core/DefaultProxy',
         'Core/loadImage',
         'Core/Math',
@@ -13,7 +13,7 @@ defineSuite([
         'Scene/UrlTemplateImageryProvider',
         'Specs/pollToPromise'
     ], function(
-        createOpenStreetMapImageryProvider,
+        OpenStreetMapImageryProvider,
         DefaultProxy,
         loadImage,
         CesiumMath,
@@ -37,12 +37,12 @@ defineSuite([
     });
 
     it('return a UrlTemplateImageryProvider', function() {
-        var provider = createOpenStreetMapImageryProvider();
+        var provider = new OpenStreetMapImageryProvider();
         expect(provider).toBeInstanceOf(UrlTemplateImageryProvider);
     });
 
     it('returns valid value for hasAlphaChannel', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server/'
         });
 
@@ -58,7 +58,7 @@ defineSuite([
             url : 'made/up/osm/server/'
         });
 
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : resource
         });
 
@@ -80,7 +80,7 @@ defineSuite([
     });
 
     it('supports a slash at the end of the URL', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server/'
         });
 
@@ -102,7 +102,7 @@ defineSuite([
     });
 
     it('supports no slash at the endof the URL', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server'
         });
 
@@ -124,7 +124,7 @@ defineSuite([
     });
 
     it('requestImage returns a promise for an image and loads it for cross-origin use', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server/'
         });
 
@@ -152,14 +152,14 @@ defineSuite([
     });
 
     it('when no credit is supplied, a default one is used', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server'
         });
         expect(provider.credit).toBeDefined();
     });
 
     it('turns the supplied credit into a logo', function() {
-        var providerWithCredit = createOpenStreetMapImageryProvider({
+        var providerWithCredit = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server',
             credit : 'Thanks to our awesome made up source of this imagery!'
         });
@@ -168,7 +168,7 @@ defineSuite([
 
     it('routes requests through a proxy if one is specified', function() {
         var proxy = new DefaultProxy('/proxy/');
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server',
             proxy : proxy
         });
@@ -194,7 +194,7 @@ defineSuite([
 
     it('rectangle passed to constructor does not affect tile numbering', function() {
         var rectangle = new Rectangle(0.1, 0.2, 0.3, 0.4);
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server',
             rectangle : rectangle
         });
@@ -227,7 +227,7 @@ defineSuite([
     });
 
     it('uses maximumLevel passed to constructor', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server',
             maximumLevel : 5
         });
@@ -235,7 +235,7 @@ defineSuite([
     });
 
     it('uses minimumLevel passed to constructor', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server',
             minimumLevel : 1
         });
@@ -243,7 +243,7 @@ defineSuite([
     });
 
     it('raises error event when image cannot be loaded', function() {
-        var provider = createOpenStreetMapImageryProvider({
+        var provider = new OpenStreetMapImageryProvider({
             url : 'made/up/osm/server'
         });
 
@@ -295,7 +295,7 @@ defineSuite([
         var rectangle = new Rectangle(0.0, 0.0, CesiumMath.toRadians(1.0), CesiumMath.toRadians(1.0));
 
         expect(function() {
-            return createOpenStreetMapImageryProvider({
+            return new OpenStreetMapImageryProvider({
                 minimumLevel : 9,
                 rectangle : rectangle
             });