+++ Specs/Scene/TileMapServiceImageryProviderSpec.js
@@ -1,5 +1,5 @@
 defineSuite([
-        'Scene/createTileMapServiceImageryProvider',
+        'Scene/TileMapServiceImageryProvider',
         'Core/Cartesian2',
         'Core/Cartographic',
         'Core/DefaultProxy',
@@ -21,7 +21,7 @@ defineSuite([
         'Specs/pollToPromise',
         'ThirdParty/when'
     ], function(
-        createTileMapServiceImageryProvider,
+        TileMapServiceImageryProvider,
         Cartesian2,
         Cartographic,
         DefaultProxy,
@@ -54,14 +54,14 @@ defineSuite([
     });
 
     it('return a UrlTemplateImageryProvider', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
           url: 'made/up/tms/server/'
         });
         expect(provider).toBeInstanceOf(UrlTemplateImageryProvider);
     });
 
     it('resolves readyPromise', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server/'
         });
 
@@ -76,7 +76,7 @@ defineSuite([
             url: 'made/up/tms/server/'
         });
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : resource
         });
 
@@ -110,7 +110,7 @@ defineSuite([
             }, 1);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -145,7 +145,7 @@ defineSuite([
             }, 1);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -159,13 +159,13 @@ defineSuite([
 
     it('requires the url to be specified', function() {
         function createWithoutUrl() {
-            return createTileMapServiceImageryProvider({});
+            return new TileMapServiceImageryProvider({});
         }
         expect(createWithoutUrl).toThrowDeveloperError();
     });
 
     it('returns valid value for hasAlphaChannel', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server/'
         });
 
@@ -178,7 +178,7 @@ defineSuite([
 
     it('supports a slash at the end of the URL', function() {
         var baseUrl = 'made/up/tms/server/';
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : baseUrl
         });
 
@@ -200,7 +200,7 @@ defineSuite([
     });
 
     it('supports no slash at the endof the URL', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'http://made/up/tms/server'
         });
 
@@ -223,7 +223,7 @@ defineSuite([
 
     it('supports a query string at the end of the URL', function() {
         var baseUrl = 'made/up/tms/server/';
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : baseUrl + '?a=some&b=query'
         });
 
@@ -245,7 +245,7 @@ defineSuite([
     });
 
     it('requestImage returns a promise for an image and loads it for cross-origin use', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server/'
         });
 
@@ -272,7 +272,7 @@ defineSuite([
     });
 
     it('when no credit is supplied, the provider has no logo', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
         return pollToPromise(function() {
@@ -283,7 +283,7 @@ defineSuite([
     });
 
     it('turns the supplied credit into a logo', function() {
-        var providerWithCredit = createTileMapServiceImageryProvider({
+        var providerWithCredit = new TileMapServiceImageryProvider({
             url : 'made/up/gms/server',
             credit : 'Thanks to our awesome made up source of this imagery!'
         });
@@ -303,7 +303,7 @@ defineSuite([
             deferred.reject(); //since the TMS server doesn't exist (and doesn't need too) we can just reject here.
         });
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'server.invalid',
             proxy : proxy
         });
@@ -322,7 +322,7 @@ defineSuite([
             deferred.reject(); //since the TMS server doesn't exist (and doesn't need too) we can just reject here.
         });
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'http://server.invalid?query=1'
         });
 
@@ -334,7 +334,7 @@ defineSuite([
 
     it('routes tile requests through a proxy if one is specified', function() {
         var proxy = new DefaultProxy('/proxy/');
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server',
             proxy : proxy
         });
@@ -360,7 +360,7 @@ defineSuite([
 
     it('rectangle passed to constructor does not affect tile numbering', function() {
         var rectangle = new Rectangle(0.1, 0.2, 0.3, 0.4);
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server',
             rectangle : rectangle
         });
@@ -393,7 +393,7 @@ defineSuite([
     });
 
     it('uses maximumLevel passed to constructor', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server',
             maximumLevel : 5
         });
@@ -406,7 +406,7 @@ defineSuite([
     });
 
     it('raises error event when image cannot be loaded', function() {
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -473,7 +473,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -511,7 +511,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -543,7 +543,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -575,7 +575,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -607,7 +607,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -648,7 +648,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });
 
@@ -688,7 +688,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server',
             flipXY : true
         });
@@ -729,7 +729,7 @@ defineSuite([
             deferred.resolve(xml);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server',
             flipXY : true
         });
@@ -774,7 +774,7 @@ defineSuite([
             }, 1);
         };
 
-        var provider = createTileMapServiceImageryProvider({
+        var provider = new TileMapServiceImageryProvider({
             url : 'made/up/tms/server'
         });